Update Playwright configuration and improve viewport type handling
- Modified the Playwright configuration to dynamically set the number of workers based on the PLAYWRIGHT_WORKERS environment variable.
- Enhanced the web server command to conditionally include coverage based on the COVERAGE environment variable.
- Updated the .gitignore file to exclude Watchman cookies.
- Refactored the getViewport method in the rendering engine to support generic types, improving type safety across viewport retrievals.
- Adjusted various viewport examples to utilize the updated getViewport method for better type inference.
- Introduced a new VolumeViewport3DLegacyAdapter to support legacy volume rendering, ensuring backward compatibility with existing volume data handling.
- Added a RenderModesPanel to the stats overlay for improved visibility of active render modes in the viewport.

+++ packages/core/src/RenderingEngine/ViewportNext/Volume3D/VtkGeometry3DRenderPath.ts
@@ -185,7 +185,7 @@ function createActorEntries(geometry: IGeometry) {
     polyData.getPoints().setData(surface.points, 3);
     polyData.setPolys(
       vtkCellArray.newInstance({
-        values: Float32Array.from(surface.polys),
+        values: Uint32Array.from(surface.polys),
       })
     );
